requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
Pillow>=10.0
mss>=9.0
opencv-python>=4.9
pytesseract>=0.3.10
PyQt6>=6.6
pyyaml>=6.0
numpy>=1.26
